Changed Model namespace
Tried to implement IAmenity. Ended up changing namespaces
