+++ Common/LoginResponse.cs
@@ -1,4 +1,4 @@
-﻿using SocietyManagementShowcase.Models;
+﻿using Common.Models;
 
 namespace Common
 {
