+++ Common/Models/CommonAmenities.cs
@@ -1,14 +1,15 @@
-﻿using SocietyManagementShowcase.Common;
+﻿using Common.Common;
 
-namespace SocietyManagementShowcase.Models
+namespace Common.Models
 {
     public class CommonAmenities
     {
         public int Id { get; set; }
         public StatusHealth Health { get; set; }
-        public List<MaintenanceLog> CommonAmenitiesMaintenaceLog { get; set; }
+        public List<MaintenanceLog> MaintenaceLog { get; set; }
         public DateTime LastMaintenanceCheck { get; set; }
         public CommonAmenitiesType Type { get; set; }
+        public AmenityType AmenityType { get; set; }
     }
 
     public enum CommonAmenitiesType
