+++ Common/Models/SwimmingPool.cs
@@ -1,14 +1,16 @@
-﻿using SocietyManagementShowcase.Common;
+﻿using Common.Common;
 
-namespace SocietyManagementShowcase.Models
+namespace Common.Models
 {
     public class SwimmingPool
     {
         public int Id { get; set; }
         public StatusHealth Health { get; set; }
-        public List<MaintenanceLog> SwimmingPoolMaintenaceLog { get; set; }
+        public List<MaintenanceLog> MaintenaceLog { get; set; }
         public DateTime LastMaintenanceCheck { get; set; }
         public SwimmingPoolType PoolType { get; set; }
+        public AmenityType AmenityType { get; set; }
+
     }
 
     public enum SwimmingPoolType
