+++ SocietyManagementUI/Filters/ManagerAuthorizationFilter.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
-using SocietyManagementShowcase.Models;
+using Common.Models;
 using SocietyManagementUI.Common;
 
 namespace SocietyManagementUI.Filters
