+++ SocietyManagementUI/Views/Amenities/AddMaintenanceLog.cshtml
@@ -1,5 +1,5 @@
-﻿@using SocietyManagementShowcase.Models
-@model SocietyManagementShowcase.Models.MaintenanceLog
+﻿@using Common.Models
+@model Common.Models.MaintenanceLog
 
 <nav>
 	<div class="nav-wrapper red accent-3 z-depth-2">
