+++ SocietyManagementUI/Views/Amenities/EditMaintenanceLog.cshtml
@@ -1,4 +1,4 @@
-﻿@model SocietyManagementShowcase.Models.MaintenanceLog;
+﻿@model Common.Models.MaintenanceLog;
 
 <nav>
 	<div class="nav-wrapper red accent-3 z-depth-2">
