+++ SocietyManagementUI/Views/Amenities/GenericMaintenanceLog.cshtml
@@ -1,4 +1,4 @@
-﻿@using SocietyManagementShowcase.Models;
+﻿@using Common.Models;
 @using SocietyManagementUI.Common
 @model List<MaintenanceLog>;
