+++ SocietyManagementUI/Views/Amenities/GymAction.cshtml
@@ -1,6 +1,6 @@
-﻿@using SocietyManagementShowcase.Models
+﻿@using Common.Models
 @using SocietyManagementUI.Common
-@model SocietyManagementShowcase.Models.Gym;
+@model Common.Models.Gym;
 
 <script type="text/javascript" language="javascript">
 	$(document).ready(function() {
