+++ SocietyManagementUI/Views/Amenities/GymActionEdit.cshtml
@@ -1,4 +1,4 @@
-﻿@model SocietyManagementShowcase.Models.Gym;
+﻿@model Common.Models.Gym;
 
 <script type="text/javascript" language="javascript">
 	$(document).ready(function() {
