+++ SocietyManagementUI/Views/Amenities/SwimmingPoolAction.cshtml
@@ -1,4 +1,4 @@
-﻿@model SocietyManagementShowcase.Models.SwimmingPool;
+﻿@model Common.Models.SwimmingPool;
 
 <nav>
     <div class="nav-wrapper red accent-3 z-depth-2">
