+++ SocietyManagementUI/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model SocietyManagementShowcase.Models.User;
+﻿@model Common.Models.User;
 @{
 	ViewData["Title"] = "Home Page";
 }
