+++ SocietyManagementUI/Views/User/AddUser.cshtml
@@ -1,4 +1,4 @@
-﻿@model SocietyManagementShowcase.Models.User;
+﻿@model Common.Models.User;
 
 <nav>
 	<div class="nav-wrapper">
