+++ SocietyManagementUI/Views/User/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@using SocietyManagementShowcase.Models;
+﻿@using Common.Models;
 @using SocietyManagementUI.Common
 @model List<User>;
